chore: Update PostgreSQL Client Authentication Configuration
The PostgreSQL Client Authentication Configuration file `pg_hba.conf` has been updated to include additional rules for replication setup and Docker internal network connections. This ensures proper authentication and access control for different IP addresses and subnets.

Refactor services in docker-compose.yml for consistency

Rename the services in the `docker-compose.yml` file to follow a consistent naming convention. The services `FrontEnd`, `BackEnd`, and `FullStack` have been renamed to `frontend`, `backend`, and `fullstack`, respectively. This improves the clarity and maintainability of the Docker Compose configuration.

Add haproxy service for load balancing

Add a new service, `haproxy`, to the `docker-compose.yml` file. The `haproxy` service uses the latest version of the HAProxy image and is responsible for load balancing the connections to the PostgreSQL database. This improves the scalability and reliability of the PostgreSQL database by distributing the incoming connections across multiple instances.

fix: readme